QDoc: Deprecate '\instantiates' command
With the introduction of the '\nativetype' command, QDoc now has an
improved mechanism for documenting what a QML type is called in
C++. The new command was introduced with the explicit aim to replace
the old, and thus this patch deprecates the latter and replaces its
use in QDoc's documentation and tests.

The old command is kept for compatibility reasons, but now delegates
to the new command internally. This means QDoc will generate the same
output for both commands until the now deprecated '\instantiates'
command is removed.

Modify the QDoc manual accordingly, and promote `\nativetype` where
suitable.

[ChangeLog][QDoc][Deprecation] The '\instantiates' command is
deprecated and will be removed in a future version of QDoc. Use the
new '\nativetype' command instead.

Task-number: QTBUG-125368
Change-Id: Ia14ef9636ec45c6c5d71d739342963b903f2c557
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>