QDoc: Replace qmlElement with qmlNativeType in ClassNode
Since the introduction of the new '\nativetype' command and the
subsequent deprecation of the '\instantiates' command, the notion of a
"qml element" that belongs to a `ClassNode` is made redundant. The only
external use of the related API in `ClassNode` is in
`Tree::resolveCppToQmlLinks`. This change replaces the call to
`setQmlElement` with a call to `insertQmlNativeType`, which is the
corresponding setter for the new QDoc command. This ensures full
feature parity between the new and old commands, and leaves the
setter, getter, and data member in `ClassNode` unused, and therefore
removed.

Task-number: QTBUG-125368
Change-Id: If0b541844f27fe4455ad1ae27c82d471611e2e81
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>