qdoc: Disambiguate identically-named C++ types in generated lists
Multiple submodules of qt5 now declare clashing C++ type names.
QDoc was capable of displaying the fully qualified name for types
under a namespace, but this is not the case here.

Further, as these types of clashes are more common in the QML side,
QDoc adds a module name suffix if there are multiple occurrences of a QML type name. Remove the check for QML types from that code, so that it applies also to C++ classes. This eliminates identical-looking links from the 'All C++ Classes' page.

Pick-to: 6.7
Fixes: QTBUG-125759
Change-Id: I7f9c043d8bd6df45cbeef5b6fd8bd307de9cfa51
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>