qdoc: Write \target and \keyword elements to index for all node types
Since the introduction of QDocIndexFiles::writeFunctionSection(), QDoc
inadvertently lost the ability to link to custom targets defined within
a \fn topic.

Reintroduce that ability by extracting the relevant logic into a helper
function and calling it also in writeFunctionSection(). This way,
targets and keywords are handled for all node types.

Pick-to: 6.5 6.7
Fixes: QTBUG-125969
Change-Id: Ifce07cf25ceb1250a33150641f04fb2736495004
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>