QDoc: Disable ASan sigaltstack in output test
Due to ASan's use of sigaltstack, it can crash with deallocation
errors if instrumenting binaries that also use sigaltstack (directly
or transitively). As LLVM has its own sigaltstack, this has caused
integration failures in COIN (see related Jira issues).

This is a known issue, see e.g:
https://github.com/google/sanitizers/issues/849
https://github.com/KDE/kdevelop/commit/e306f3e39aba37b606dadba195fa5b7b73816f8f

Modify the test binary such that it sets or extends the `ASAN_OPTIONS`
environment variable for QDoc's process environment, to avoid crash in
ASan. Promote the existing QProcessEnvironment to a static that is
shared between the `initTestCase` and `runQDocProcess` member
functions, to avoid the cost of repeatedly creating the QPE object.

Fixes: QTBUG-126116
Task-number: QTBUG-111580
Pick-to: 6.8
Change-Id: I25c7baed9ccec14aaad0c47512b58118a3a91d84
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>