QDoc: Add test project for compiler generated member functions
QDoc can generate documentation for compiler generated member
functions. This change adds a test project that exercises this
capability.

QDoc requires the full signature as generated by the compiler as an
argument to the `\fn`-command. Note, however, that for an
`= default`ed member function, it is not required to specify
`=default` in the `\fn` argument.

Task-number: QTBUG-123130
Change-Id: Iff52d827c48aa14bdaeefae53387ccf5230802b4
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>