Qt Designer/property browser: Set editor background to white in case of transparent background color
The property editors inherit the QPalette of the underlying
QAbstractItemView::viewport. In case of the QWindows11Style, this color
corresponds to Qt::transparent for the backgroundRole. This leads to the  original column of the QTreeView not being painted over by the editor and it may happen that the value is visible two times during editing. This patch sets the color of the editor explicitly to white in case of a
transparent background, such that the original content is again painted over.

Fixes: QTBUG-126274
Change-Id: I4434a738402cf7dc34a2f1b551d5c1031acc65a9
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>
(cherry picked from commit 9d660b4fd454d8d47371f2ed52d24b5b8a9e157e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit 9a137fbd78a93570489b630a03d7db58680e2603)