qdoc: Remove unnecessary test when merging QML module nodes
Some documented QML modules, such as QtQuick.LocalStorage, do not
contain any member types. In case of LocalStorage, it exposes a
singleton type whose methods are documented directly on the
module page.

QDocDatabase::mergeCollections() implementation included a check for
whether the collection's list of members was empty. If it was, the
node was not included in the generated output e.g. via

  \generatelist modules-qml

Drop this check and requirement - allow generated lists to include
"empty" but documented modules.

Pick-to: 6.7 6.8
Fixes: QTBUG-61277
Change-Id: Ia2106cb11b32df10bf4d559edfd8c2469fdbd9da
Reviewed-by: Paul Wicking <paul.wicking@qt.io>