qdoc: Fix Aggregate::nonfunctionList()
A common way to remove duplicates from a container is to
first sort it, then use std::erase() together with std::unique()
on the container.

However, Aggregate::nonfunctionList() applied a custom sort
function when doing this; this meant that duplicate pointers
in the container were not (necessarily) adjacent to each other
after sorting, as the sorting is based on other attributes.

To fix, sort twice - once based on pointer value, final sorting
by name.

Add a test in Node::nodeNameLessThan() to compare two identically-
named QML types based on their module name. This ensures correct
alphabetical sorting of QML types.

Pick-to: 6.9 6.8 6.5
Change-Id: I80df5ab0d77f39f104bef434aa84f8260caece3b
Reviewed-by: Paul Wicking <paul.wicking@qt.io>