CMake: Disable lupdate/clang if llvm was built without RTTI support
The linking of the clang-based lupdate fails if llvm was built without
RTTI support. Add this condition to the clangcpp feature that controls
the clang-based lupdate.

Also, extend the clangcpp warning message like so:

WARNING: LLVM was found, but it was not built with RTTI support. The
Clang-based lupdate parser will not be available. Suitable LLVM and
Clang C++ libraries have not been found. You will need to set the
FEATURE_clangcpp CMake variable to ON to re-evaluate this check.

If LLVM was built with RTTI but is too old or otherwise unsuitable, the
first sentence is missing.

Pick-to: 6.8
Fixes: QTBUG-101446
Change-Id: I3d0991763736104a231de7cf82b6f282e6e74e9d
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>