qdoc: Warn if qhp subproject indexTitle is not found
QDoc warned about a missing index title for a qhp subproject, but only
if the subproject.type was 'manual'.

In addition, the warning that was generated was not detailed enough, as
it was created using an incorrect Location instance that did not
specify the problematic file (in this case, .qdocconf).

Adjust the warning text, and ensure that it is generated for all
subproject types. Add documentation, and remove incorrect documentation
entry for the old warning.

Pick-to: 6.8
Fixes: QTBUG-126960
Change-Id: Ied37244749db4645acf2819a9571b3a14e71ca2d
Reviewed-by: Paul Wicking <paul.wicking@qt.io>