lupdate: Make exclude patterns less strict again
TR_EXCLUDE and QT_EXCLUDE_SOURCES_FROM_TRANSLATION became too strict
after the code base was ported from QRegExp to QRegularExpression in
commit 501186daee80601d77d21d590a9fa3d6217cfbe8.

Before that change, a pattern like
    TR_EXCLUDE += /foo/bar/*
would exclude files in the directory /foo/bar and all subdirectories.

After the port to QRegularExpression, only files in the /foo/bar
directory were excluded because
QRegularExpression::wildcardToRegularExpression creates an anchored
expression by default.

Now, we centralize the creation of the regular expressions and create
expressions that are anchored at the start only such that files in
subdirectories can be excluded again.

Pick-to: 6.5 6.7 6.8
Fixes: QTBUG-127146
Change-Id: Ia2eff6d19df61ccd451e529c63c6ba40c63546b0
Reviewed-by: Kai Köhne <kai.koehne@qt.io>