lupdate: Fix long runtime with supposed includes in namespaces
Lupdate was extremely slow for projects with certain libboost includes.
The reason is that the boost includes define types in the boost
namespace mixed with pre-processor conditionals such that lupdate
detected further boost includes within namespaces. See the linked issue
for details.

Those included files were then exempt from any caching. That was not
just a performance penalty, but a working include file cache is also
needed for breaking out of include cycles.

The following condition was preventing the usage of the cached parsing
result:
    if (namespaces.size() == 1 && functionContext.size() == 1
        && functionContextUnresolved.isEmpty()
        && pendingContext.isEmpty() ...)
        // then use the parse result
These four members are exactly the CppParserState.

We now not only use the file path as key for the parser results cache,
but the parser state in addition. This allows us to remove the parser
state from the condition above and still return correct parser results.

Fixes: QTBUG-27936
Pick-to: 6.5 6.7 6.8
Change-Id: I9d7c22ad04fa42d39778ee6637c9cbbfd681d9bf
Reviewed-by: Kai Köhne <kai.koehne@qt.io>