qdoc: Allow -DDEFINE=VALUE command line options
QDoc supported setting macros from the command line with the -D option,
however, it required whitespace between the option and its value.

Set the QCommandLineOption::ShortOptionStyle flag for the option, making
it behave identical to -I, -isystem, and -F options, and in line with
how these options are passed to compilers from the command line.

Pick-to: 6.8
Fixes: QTBUG-83969
Change-Id: I10a6af5249086870ab12366a89bdce8ba26b9fb7
Reviewed-by: Paul Wicking <paul.wicking@qt.io>