qdoc: Fix the qhp group selector for unseen groups
QDoc supports grouping items/pages with the \ingroup command, and lists
of group members can be generated with the \generatelist <group_name>
command. This works even if the group is not documented with the
\group command.

Likewise, the qhp subproject definition in .qdocconf supports the
group:<group_name> selector. But here, the selector worked only if the
group is documented, that is, a \group page exists for it.

The reason was that undocumented group nodes (a.k.a unseen groups) were
marked not only 'unseen' but also internal, as they themselves are
omitted when generating documentation.

Add a special exception for unseen groups in HelpProjectWriter, so that
their members are correctly added to the .qhp if they otherwise match the defined selector.

Pick-to: 6.8
Fixes: QTBUG-127564
Change-Id: Id06aa4169fd6d9b2ca828cb7929df47761068cf3
Reviewed-by: Paul Wicking <paul.wicking@qt.io>