QDoc: Format parts of requisites table as code in HTML output
In API documentation for C++ and QML, QDoc generates a `requisites`
table that lists information about, for example, the header to include
to gain access to a class or namespace, CMake instructions, etc. As
the intention is to provide this as copy-pasta, make the information
stand out visually by wrapping it in the `code` html element, to
ensure it's rendered with a monospace font face.

Extend the CSS class for the table with the new identifier
`requisites`, such that the table can be styled with ease.

Fixes: QTBUG-127617
Change-Id: I78f31317feb2245e69bb2703484dcc95318627a9
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
(cherry picked from commit eb3d903cffa4d1610812a86e201a774f4430ef4d)