QDoc: Allow author to specify `auto` as return type in `\fn` command
When documenting a C++ function using the `\fn` command, QDoc expects
the fully qualified type as part of the signature passed to that
command. If a function is declared and defined in source using the
`auto` keyword to let the compiler deduce the type, QDoc correctly
generates `auto` as the type for the function in most cases. However,
if the type is a template specialization, QDoc's `ClangCodeParser`
will set the type to the compiler deduced type instead of
`auto`. Though technically correct, this may depart from the author's
intention, and may expose implementation details of an API that are
meant to be hidden. In the case of Qt, this may mean exposing the
`QtPrivate` namespace, which client code shouldn't use directly.

This patch seeks to resolve this issue and thus allow an author to
specify `auto` as the return type passed to QDoc's `\fn` command:

* Add a test project that makes use of `auto` as return type in
  documentation.
* Change the type of `FunctionNode::m_returnType` from `QString` to
  `QPair<QString, std::optional<QString>>`, such that it may hold an
  optional second value.
* Update `FunctionNode::setReturnType` and `FunctionNode::returnsBool`
  to operate on the first element of `FunctionNode::m_returnType`.
* Add `FunctionNode::declaredReturnType` as a getter for the second
  element of `FunctionNode::m_returnType`.
* Add `FunctionNode::returnTypeString` as a convenience getter for the
  return type, which returns the second element of
  `FunctionNode::m_returnType` if it is set, otherwise the first
  element.
* Modify `ClangCodeParser::visitFnSignature` to set the declared
  return type on the `FunctionNode` if the node's return type differs
  from that which is declared in the `\fn`-command, and if the
  declared return type is `auto`.
* Modify `CppCodeMarker::markedUpSynopsis` and such that it calls the new
  convenience getter for the return type string. Make the same changes
  to call sites in `DocBookGenerator`.

[ChangeLog][QDoc] QDoc now respects the author's intent better, in that it outputs `auto` as the return type of functions that are documented with `auto` as their return type.

Fixes: QTBUG-126537
Change-Id: I5fa18effca2ee69937a2d66763213948c09e5e8c
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
(cherry picked from commit d31d4735e21d5a8338c305b8e801a07572a93db6)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>