CMake: store 'language' and 'sourcelanguage' in generated .ts files
The initially created .ts files were lacking the 'language' and
'sourcelanguage' attributes. We did not write them, because the first
time the file is opened in Qt Linguist, a dialog will request the user
to set these languages. However, without a proper 'language' attribute,
lupdate is unable to generate the right amount of plural forms on "ninja
update_translations".

We now store the translated language as source file property when
calculating the file path and write it to the initially generated .ts
file. For 'sourcelanguage' we simply write the value of
QT_I18N_SOURCE_LANGUAGE.

Pick-to: 6.7 6.8
Fixes: QTBUG-127789
Fixes: QTBUG-127792
Change-Id: Ib5eca421233f9ab26bf2eaf6781269baa0f5dbed
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>