lupdate: Add architecture condition for compiler flag passed to Clang
lupdate passes options to Clang when it calls the compiler. Among
these options is the flag `-mno-sse`. This flag is intended to disable
the generation of SIMD instructions for SSE, and is applicable to x86
and x86_64 architectures. The option doesn't exist for other
architectures. When lupdate is linked against `libclang` built for
ARM(64), its compiler invocation thus fails.

Wrap the compiler option in a check for the underlying system
architecture such that it is passed to clang only if the system
architecture is a match. Exclude systems running macOS from this check,
as macOS provides Rosetta 2 which provides a way to run X86 code on
ARM. A comment is modified for added precision, and its contents
reflowed to stay within conventional column bounds.

Task-number: QTBUG-127751
Pick-to: 6.8
Change-Id: If91d9b5fa45d5836953d9ea8bff9670851caff06
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>