Doc: Remove usage of defunct QDoc command
The QDoc command `\tableofcontents` was rendered useless by a change
more than a decade ago. Remove the use of `\tableofcontents` as it
serves no purpose, and ensure that the surrounding context still makes
sense for the reader, by removing preceding text that refers to a
non-existing table of contents, such as `Contents:`.

Task-number: QTBUG-128173
Pick-to: 6.8
Change-Id: I9187d7b9da88c431f0eac33922bb1a2b3499a0f6
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
Reviewed-by: Safiyyah Moosa <safiyyah.moosa@qt.io>