QDoc manual: Remove dud "Core Features" example page
The QDoc manual contains a code snippet that shows how to employ the
`\include` command. This snippet uses the command in a mock "Core
Features" page. The documentation goes on to provide a link to the
file as it would be generated by QDoc, to provide a visual example.

Both the file name and page title are problematic, as there's nothing
that indicates that this is an example and not a page that should
describe QDoc's core features. Furthermore, as this page is indexed by
search engines, it may be presented in a prominent position in search
results. This includes, and is likely not limited to, searching for
the string "core features" in the search box on doc.qt.io.

Remove the file and the line that references it, as it is deemed
unnecessary.

Fixes: QTBUG-128411
Change-Id: Ie7d7949588f4b1f93fd55a4e4467d9f98a7a6c75
Reviewed-by: Kai Köhne <kai.koehne@qt.io>
(cherry picked from commit a2dcb23e55551330f92c9c56d4db6a07331c2209)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit e170d9894b97225906e090d18c5fed5cab3a7986)