Doc: Don't claim that lupdate "does not destroy any data"
There was this bit in the docs:
"lupdate synchronizes the user-visible text from the application with
the translations. It does not destroy any data."

This is a bit misleading. Added XML comments, or custom tags, or
untranslated messages are not preserved when lupdate is run on a .ts
file.

Let's be a bit less bold and just say that lupdate preserves existing
translations.

Fixes: QTBUG-52024
Change-Id: Ide23c18b38193507095fc95ef9af6d7ec8ebaf9f
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>
Reviewed-by: Paul Wicking <paul.wicking@qt.io>