Doc: Fix wrong command name in QDoc manual
While bulk changing use of the `\instantiates` command to the new
`nativetype` command, the entries in the QDoc manual for
`\instantiates` were changed on accident. This lead to a confusing
statement in the documentation:
  "The \nativetype command is deprecated since Qt 6.8.
   Use \nativetype instead".

This change corrects the command name in the documentation.

Task-number: QTBUG-128216
Change-Id: Ib1f42d2166c1ccfb231885ba21bf35c3d0752194
Reviewed-by: Andreas Eliasson <andreas.eliasson@qt.io>
(cherry picked from commit 3a0419d51a5cfa32b02ecf5ff156ddaf2eb1cb02)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>