qdoc: Allow tying a \keyword and a \target to a section title
Previously, a \keyword could only be used to create a link target to
the topic it appears in. Adding a link target to a specific location
within the topic (e.g. a \page) was only possible with the \target
command; however, a \target does not add a keyword entry to the
Qt Help project (.qhp) file like a \keyword does.

Enable tying a \keyword to a \section1..4 title within a documentation
comment. If a \keyword appears directly before a \section command, the
section's href is adopted by the keyword and recorded into the .qhp.
This way, it's possible to have keyword in the Assistant's or Qt Creator
Help's index that refer to a section. These indexed sections are also
available for the context help.

Likewise, these changes alter the behavior of the \target command to be
slightly cleaner:

  \target hola
  \section1 Hello

Will make \l {hola} link to the anchor generated for the section,
instead of a (somewhat arbitrary) location just above the section.

Add a new value `ContentsKeyword` to TargetRec::TargetType enum to
facilitate this. As a drive-by, add internal documentation for all
TargetType enum values to clarify their roles.

Fixes: QTBUG-128573
Change-Id: Ifacdfab174f1aad7fcb7420416e7b56c461ea04e
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
(cherry picked from commit 9e94ec6a04f10fade025dcf0e359c2a2aa95e8aa)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>