CMake: Work around libclang packaging issues in FindWrapClang.cmake
The CMake package files for libclang are on Linux (and possibly elsewhere)
scattered among packages. ClangConfig.cmake is usually installed
together with clang, but required depedencies to actually use it are
packaged somewhere else. This leads to errors when doing
    find_package(Clang CONFIG)
despite this being an optional package dependency. That means that the
configuration of qttools (or top-level Qt) fails if 'clang' has been
installed without 'libclang-<version>-dev' (Debian) or 'llvm' (Arch).

This problem has been reported to upstream:
https://github.com/llvm/llvm-project/issues/45628

We now work around this packaging issue by performing various checks
before attempting to find the 'Clang' package:
  1. find LLVM
  2. find a library that is packaged with libclang
If that is successful,
  3. find Clang with exactly the version reported by LLVM.

Users can revert to the old behavior by setting
QT_NO_FIND_PACKAGE_CLANG_WORKAROUND=ON when configuring Qt.

Fixes: QTBUG-124255
Change-Id: I1ede103fb776b6ae7c6616080e3f99409156bcbb
Reviewed-by:  Alexey Edelev <alexey.edelev@qt.io>