qdoc: Fix double entries for shared comment node in detailed description
If multiple API functions are documented directly under a namespace and
the functions share a QDoc comment block, duplicate entries for them
were generated in the detailed description - once with the correct
documentation, and a second time with empty docs.

This happened because Sections::stdRefPageSwitch() was indiscriminantly
called for both `Summary` and `Details` sections, and that function did
not check whether the node being distributed is sharing a comment (for
`Details`).

Extract the recurring piece of code that tests for shared comment nodes
as a static helper method, and use it in all call sites doing node
distribution to various sections. This makes the output for namespace
reference pages more uniform with ones for classes and QML types.

Add a validatedqdocoutputfiles test case with expected output.

Pick-to: 6.8
Fixes: QTBUG-120186
Change-Id: If9f2d4c006c19958a4948cd3dfd395d215b6c97f
Reviewed-by: Paul Wicking <paul.wicking@qt.io>