qdoc: Fix links for access functions associated with multiple properties
FunctionNode::hasOneAssociatedProperty() was used for resolving where to
link to for undocumented property access functions. If a function is
undocumented, it should link to the property where QDoc auto-generates
the list of access functions.

This broke down if there are multiple properties associated with, for
example, an undocumented notifier signal. There, QDoc gave up and just
generated a non-existent anchor for the signal, resulting in broken
intra-page links.

To fix, replace hasOneAssociatedProperty() with a function that returns
a `primary` associated property to get consistent linking. A property is
considered a primary for an access function if the function name appears
at the start of the property name.

Fixes: QTBUG-128356
Pick-to: 6.8
Change-Id: Ie23fe62bb97f3f883c794bd94042c4d376f45699
Reviewed-by: Paul Wicking <paul.wicking@qt.io>