QDoc: Add test project with C++ documentation in headers only
When generating documentation for C++ projects, QDoc considers only
comments in implementation files (e.g. files with the ".cpp"
extension). This is due to QDoc's history as being developed mainly
with the needs of the Qt Project in mind.

Other software projects, such as those developed by KDE, place
documentation comments in header files (e.g. ".h"). These projects
cannot successfully make use of QDoc as their documentation generator
without copying or moving the documentation comments, which is
considered an unreasonable cost that does not outweigh the perceived
benefits.

This patch adds a documentation project to QDoc's end-to-end test,
`tst_validateQdocOutputFiles`, which is adapted from an existing
test case. The purpose is to test QDoc's capability of generating
documentation from header file comments, in the same manner it does
for projects with documentation in implementation files. The goal is
to ascertain "feature parity" in the sense that QDoc should generate
the same output regardless of whether the underlying documentation
comment is found in a header file or an implementation file.

This test is a prerequisite for the patch that adds the documentation
from header file feature.

The test is adapted from the test case `comprehensiveproject`. Some
notable differences are worth mentioning:

- This test case does not conditionally define parts of the test
  project, and as such generates certain bits of documentation that
  are not part of the test case it's adapted from.
- Documentation is generated for both C++ and QML types, as well as
  from a `.qdoc` file, as is commonly the case in any non-trivial
  documentation project that utilizes QDoc.
- All comments in header files include QDoc's `\fn` command. This is
  due to an implementation detail in QDoc, which keeps QDoc from
  automatically tying a documentation comment to a declaration, it
  requires the full definition. This behavior is now considered a bug.

Fixes: QTBUG-129215
Task-number: QTBUG-94746
Change-Id: Ie5ad21cf5f741c2a92d61fcc2d662eca3c82b88b
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>