QDoc: Remove unused member ConfigStrings::ALIAS
The implementation of QDoc's `Config` class includes a struct,
`ConfigStrings`. This struct is a wrapper for static QStrings, and
is a Singleton-like monostate container that provides a globally
accessible, single set of configuration options across the entire
application.

This struct defines the member field `ALIAS`. However, this field is
never initialized with a default value, never set with a value read
from a `qdocconf` file, and never read during QDoc's execution. Hence,
the field is deemed unused and unnecessary, and thus removed.

Pick-to: 6.8
Change-Id: I0eb29a72b26c8c3210401780f98eb133d5257f14
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>