QDoc: Adapt to breaking changes in LLVM 19
Due to upstream changes in LLVM 19, QDoc fails to compile when linked
against Clang libraries from this version of LLVM. Three issues arise;

- Two cases of passing an argument of wrong type to a function.
- One case of accessing a non-existing member of an enumeration.

One upstream change (see [0]) is responsible for two of the issues:

- `get_expression_as_string()` is modified such that it correctly
  obtains the parameter type when calling
  `get_fully_qualified_type_name()`, by appending
  `.getArgument().getAsType()` to the call to `getDefaultArgument()`.
- `get_default_value_initializer_as_string()` is modified such that it
  correctly passes the source expression to
  `get_expression_as_string()`, by appending `.getSourceExpression()`
  to the call to `getDefaultArgument()`.

Both of these changes are is incompatible with QDoc built against
Clang libraries from earlier versions of LLVM, and are therefore
wrapped in #if-ery.

Finally, LLVM 19 drops a value used in QDoc from the enumeration
`clang::TemplateName::Qualified`, see [1]. The enum value `Fully` is
removed without replacement. The enum is left with two values,
`AsWritten` and `None`. QDoc is modified such that it relies on the
former of the two. This change doesn't cause any change in output from
QDoc when built against Clang libraries from LLVM 17 and 18, and the
change is therefore not wrapped in #if-ery.

[0] - https://github.com/llvm/llvm-project/commit/e42b799bb28815431f2c5a95f7e13fde3f1b36a1
[1] - https://github.com/llvm/llvm-project/commit/9c4a716c12920

Done-with: Khem Raj <raj.khem@gmail.com>
Fixes: QTBUG-128644
Change-Id: I34fbb46cf28b5676b4adda5e563d6d59fc40f602
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>