QDoc: Make the productname in \since command configurable
QDoc's `\since` command allows authors to specify the version in which
something is introduced to the product being documented. The command
takes a space-separated list of arguments. If the argument consists of
only one element, the assumption is that it is a Qt version, and QDoc
generates a line of text that includes "Qt" in the name. For example,

    `\since 1.2.3`

will inevitably end up as `Qt 1.2.3` in the output, regardless of the
project name or author's intent. As a consequence, non-Qt project must
therefore always include the product name as part of the command, e.g.

    `\since My Awesome Product 1.2.3.`

This causes needless repetition for authors of documentation that
doesn't belong to the Qt Project. The behavior violates the principle
of least astonishment by injecting `Qt` for some uses of the command.

This change introduces a new configuration variable to QDoc,
`productname`. The core aspect of the behavior thus remains unchanged;
if given one element only as argument, the `\since` command will still
make QDoc generate a product name as part of the output. However; the
default is no longer `Qt`; this must be reflected in the configuration
file(s) for the product(s) under documentation.

Hence, all relevant configuration files for QDoc's end-to-end tests are
modified to make use of the new configuration variable, which ensures
that the tests keep passing without modifying their expected output.

The QDoc manual is updated to reflect the new variable and new
behavior of the `\since` command.

[ChangeLog][QDoc][Behavior change] You can now configure the product
name that's used with QDoc's `\since` command with the new
configuration variable `productname`. This constitutes behavior
change, as the command previously would inject `Qt`.

Fixes: QTBUG-128911
Change-Id: I54ebcc3ee18c5b77d2d5180effebc244371949e8
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>