qdoc: Improve generation of versioned page titles
Remove code with hard-coded `Qt` inserted into the .html page
<title>, replacing it with the value from the new `productname`
.qdocconf variable.

Avoid injecting the resolved project version directly after the
home- or landing page, as in a documentation build with
multiple projects, attaching the top-level version number to
an individual subproject title might not always make sense.
Instead, if `productname` is defined, prefix the version with
the product name it's associated with.

This causes the generated <title>s to change in Qt's documentation.
For example, in the QString reference, it changes from

   QString Class | QtCore 6.9.0

to

   QString Class | QtCore | Qt 6.9.0

In passing, take `productname` also into use in the DocBook
generator.

[ChangeLog][QDoc] QDoc now uses the 'productname' configuration variable in the generated .html <title> elements.

Fixes: QTBUG-128668
Change-Id: I30de8df00cfaa24371ae2dd13bb29bb5b390e2c0
Reviewed-by: Paul Wicking <paul.wicking@qt.io>