QDoc: Double the buffer size for Tokenizer
`Tokenizer` uses a fixed-size buffer when parsing the sources.

When the buffer is filled, the parsing continues but all characters from
the currently parsed token that do not fit into the buffer are
discarded.

The limit was recently surpassed by some of the auto-generated QDoc
comment-blocks from the Squish documentation.

While the incriminated comment-blocks will be reduced in size in the
future, it was decided to increase the buffer size to allow for some
more breathing space.

Hence, the size of the buffer was doubled, to about 1Mb.

Fixes: QTBUG-128840
Done-with: Luca Di Sera <luca.disera@qt.io>
Change-Id: I0962e367470e57386bc0e15f58be979e4a1a3692
(cherry picked from commit 6e74f7b506935f18a33cd5918a1c6c4f4c9cddc4)
Reviewed-by: Paul Wicking <paul.wicking@qt.io>