qdoc: Allow auto-tying templated function docs in the header
QDoc has the capability to auto-tie a function documentation
comment to the correct declaration in the header, if the
comment block immediately precedes the function definition in
the implementation file.

With the `documentation-in-headers` feature enabled, this
process failed for comments in the header preceding a templated
function declaration. This is because when resolving the
Node representing the function, QDoc first tries to locate the
Node for the function's semantic parent. QDoc assumed this parent
to be an Aggregate (i.e, a class or a namespace), and gave up if
it's not.

In case of templated functions, the first declaration seen after
the comment is of kind template type, non-type, or template
parameter, and the semantic parent of that declaration is the
function itself, not the underlying Aggregate. As
`findNodeForCursor()` uses recursion to locate the semantic parent,
at that point we already have the correct Node instance the
documentation is referring to.

Add a test project that replicates the existing `templatedcallables`
test but with auto-tied docs in the header.

Modify existing test by removing the fn command for a template, without
modifying the expected output for that test.

Task-number: QTBUG-94746
Change-Id: I98ca42106b78ef75c88c99f39a5f6deb980ed41d
Reviewed-by: Paul Wicking <paul.wicking@qt.io>