qdoc: Allow auto-tying hidden friend functions with docs in the header
Hidden friend functions are recorded as children of their lexical
parent in the QDoc database. With the `documentation-in-headers`
feature enabled, when parsing the declaration of a hidden friend,
Clang sees it under its semantic parent. Therefore, in order to
tie the documentation to the declaration, an \fn command with
a fully qualified path (including the lexical parent scope) was
required.

To allow matching the doc with the declaration without an \fn
command, detect the friend declaration for functions in
`findNodeForCursor()` and retry searching for them in the database
under their lexical parent.

Task-number: QTBUG-129289
Change-Id: Ia41756c01276ad00b926906a1fc4f3bb8a732c97
Reviewed-by: Paul Wicking <paul.wicking@qt.io>