Qt Designer: Do not show `New Form` window when client or server mode
When client or server mode is used, the `New Form` window is not
needed because a file is sent to the designer just after running
Designer. If we don't hide the window, both the `New Form` window and
the file is shown at the same time.

Change-Id: I98bc61a6fbce5808e268598e057e02fad21f749b
(cherry picked from commit fde68d6d5bb8a2ea7566b775a66eb9053952a6b3)
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>