qtattributionsscanner: Check Id value and warn
The attribution Id key JSON value is part of the argument for
the \page qdoc command, and a valid 'Id' must be lower case
and without spaces. This is documented in
[QUIP-7|https://contribute.qt-project.org/quips/7], which describes
the 'qt_attribution.json' file format. Qdoc from 6.8.1 complains
about conflicting file names for attributions with 'Id' values that
begin with the same string. For example, if there are two attributions
for 'Noto Sans Thai' and 'Noto Sans JP` fonts in the same
QDocModule, qdoc tries to generate two HTMLs with the same name,
ignoring rest of the 'Id' value after the first space.

Ideally, the tool should validate each attribution 'Id' and warn if
not valid. This change adds a check to warn and stop generating the license
attribution, if an attribution 'Id' value is not in lower case and has spaces.

Pick-to: 6.8
Change-Id: I543ae35e11d19e65f1531691ff73f4ce199f05b0
Reviewed-by: Kai Köhne <kai.koehne@qt.io>
(cherry picked from commit 8acf89cd94eaa22947787db70e58321821583e77)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>